rename: 사용자의 신고 기능 폴더 위치 수정
