+++ src/main/java/com/dtalks/dtalks/report/enums/ReportType.java
@@ -1,4 +1,4 @@
-package com.dtalks.dtalks.admin.report.enums;
+package com.dtalks.dtalks.report.enums;
 
 public enum ReportType {
     SWEAR_WORD, OTHER;
