+++ src/main/java/com/dtalks/dtalks/report/enums/ResultType.java
@@ -1,4 +1,4 @@
-package com.dtalks.dtalks.admin.report.enums;
+package com.dtalks.dtalks.report.enums;
 
 public enum ResultType {
     WAIT, SUSPENSION, BAN, NP
