+++ src/main/java/com/dtalks/dtalks/report/repository/ReportRepository.java
@@ -1,6 +1,6 @@
-package com.dtalks.dtalks.admin.report.repository;
+package com.dtalks.dtalks.report.repository;
 
-import com.dtalks.dtalks.admin.report.entity.Report;
+import com.dtalks.dtalks.report.entity.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportRepository extends JpaRepository<Report, Long> {
